+++ mobile-docs/.vitepress/theme/index.ts
@@ -2,11 +2,11 @@ import './styles/index.css'
 import VPApp from './components/VPApp.vue'
 import VPNotFound from './components/VPNotFound.vue'
 // @ts-ignore
-import '@{symbol}/delight-charts-mobile/dist/style.css'
+import '@dx/delight-charts-mobile/dist/style.css'
 import { Theme } from 'vitepress'
 import { withConfigProvider } from './composables/config'
 // @ts-ignore
-import DelightChartsMobile from '@{symbol}/delight-charts-mobile'
+import DelightChartsMobile from '@dx/delight-charts-mobile'
 
 import { vuePlugin } from "./plugins/mdx";
 import "./plugins/mdx/style.css";